+++ .trebired/logger/config.ts
@@ -1,4 +1,4 @@
 export default {
-  forVersion: "2.5.26",
+  forVersion: "2.6.3",
   prefix: "trebired.frontend",
 };